refactor: put inputs components to separate folder
