+++ src/components/pages/Home/Form/inputs/SelectInput.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { InputProps } from './NameInput';
-import { Tactility } from '../../../../types';
+import { Tactility } from '../../../../../types';
 
 class SelectInput extends Component<InputProps> {
   inputRef = React.createRef<HTMLSelectElement>();
@@ -14,7 +14,7 @@ class SelectInput extends Component<InputProps> {
       <>
         <div className="home-create-form__group select-elem">
           <div className="home-create-form__label">Tactility</div>
-          <select className="home-create-form__select" ref={this.inputRef}>
+          <select className="home-create-form__select" ref={this.inputRef} required>
             <option className="home-create-form__option">{Tactility.LoveToHug}</option>
             <option className="home-create-form__option">{Tactility.HardToGet}</option>
           </select>
